% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{basemethods}
\alias{basemethods}
\alias{basemethods,IBRAPerformance-method}
\alias{basemethods,IBRAPlot-method}
\title{Accessor function for basemethods}
\usage{
basemethods(x, ...)

\S4method{basemethods}{IBRAPerformance}(x)
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}
}
\value{
A character vector of all methods represented in the object.
}
\description{
Accessor function to extract the methods that are represented in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtprcurve")
basemethods(ibraperf)
}
\author{
Charlotte Soneson
}

