% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{corr}
\alias{corr}
\alias{corr,IBRAPerformance-method}
\alias{corr<-}
\alias{corr<-,IBRAPerformance,data.frame-method}
\title{Accessors for the 'corr' slot of an IBRAPerformance object}
\usage{
corr(x, ...)

corr(x, ...) <- value

\S4method{corr}{IBRAPerformance}(x)

\S4method{corr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'corr' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

