% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{deviation}
\alias{deviation}
\alias{deviation,IBRAPerformance-method}
\alias{deviation,IBRAPlot-method}
\alias{deviation<-}
\alias{deviation<-,IBRAPerformance,data.frame-method}
\alias{deviation<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{deviation} slot}
\usage{
deviation(x, ...)

deviation(x, ...) <- value

\S4method{deviation}{IBRAPerformance}(x)

\S4method{deviation}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information necessary to plots of deviations
between observed and true scores for each method and each stratification
level.}
}
\description{
Accessor and replacement functions for the \code{deviation} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\author{
Charlotte Soneson
}

