% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrnbr}
\alias{fdrnbr}
\alias{fdrnbr,IBRAPerformance-method}
\alias{fdrnbr,IBRAPlot-method}
\alias{fdrnbr<-}
\alias{fdrnbr<-,IBRAPerformance,data.frame-method}
\alias{fdrnbr<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{fdrnbr} slot}
\usage{
fdrnbr(x, ...)

fdrnbr(x, ...) <- value

\S4method{fdrnbr}{IBRAPerformance}(x)

\S4method{fdrnbr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information about the observed FPR and the
number of features called positive for each method and each stratification
level, at various adjusted p-value thresholds.}
}
\value{
The accessor function returns a data frame giving information about
  the observed FPR and the number of features called positive for each method
  and each stratification level, at various adjusted p-value thresholds.
}
\description{
Accessor and replacement functions for the \code{fdrnbr} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrnbr")
head(fdrnbr(ibraperf))
}
\author{
Charlotte Soneson
}

