% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrtpr}
\alias{fdrtpr}
\alias{fdrtpr,IBRAPerformance-method}
\alias{fdrtpr,IBRAPlot-method}
\alias{fdrtpr<-}
\alias{fdrtpr<-,IBRAPerformance,data.frame-method}
\alias{fdrtpr<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'fdrtpr' slot of an IBRAPerformance object}
\usage{
fdrtpr(x, ...)

fdrtpr(x, ...) <- value

\S4method{fdrtpr}{IBRAPerformance}(x)

\S4method{fdrtpr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'fdrtpr' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

