% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrtprcurve}
\alias{fdrtprcurve}
\alias{fdrtprcurve,IBRAPerformance-method}
\alias{fdrtprcurve,IBRAPlot-method}
\alias{fdrtprcurve<-}
\alias{fdrtprcurve<-,IBRAPerformance,data.frame-method}
\alias{fdrtprcurve<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{fdrtprcurve} slot}
\usage{
fdrtprcurve(x, ...)

fdrtprcurve(x, ...) <- value

\S4method{fdrtprcurve}{IBRAPerformance}(x)

\S4method{fdrtprcurve}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information necessary to generate curves of
observed FDR vs TPR for each method and each stratification level.}
}
\value{
The accessor function returns a data frame giving information
  necessary to generate curves of observed FDR vs TPR for each method and
  each stratification level.
}
\description{
Accessor and replacement functions for the \code{fdrtprcurve} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtprcurve")
head(fdrtprcurve(ibraperf))
}
\author{
Charlotte Soneson
}

