% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fpr}
\alias{fpr}
\alias{fpr,IBRAPerformance-method}
\alias{fpr,IBRAPlot-method}
\alias{fpr<-}
\alias{fpr<-,IBRAPerformance,data.frame-method}
\alias{fpr<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{fpr} slot}
\usage{
fpr(x, ...)

fpr(x, ...) <- value

\S4method{fpr}{IBRAPerformance}(x)

\S4method{fpr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information about the observed FPR for each
method and each stratification level, at various adjusted p-value
thresholds.}
}
\value{
The accessor function returns a data frame giving information about
  the observed FPR for each method and each stratification level, at various
  adjusted p-value thresholds.
}
\description{
Accessor and replacement functions for the \code{fpr} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fpr")
head(fpr(ibraperf))
}
\author{
Charlotte Soneson
}

