% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAData.R
\docType{class}
\name{IBRAData-class}
\alias{.IBRAData}
\alias{IBRAData-class}
\title{A class to hold observed scores and true values for features}
\description{
The IBRAData class contains slots to hold calculated p-values, adjusted
p-values and general 'scores' for a set of features. The slots can contain
values from multiple methods, and each method can contribute to one or more
slots. The class also contains a slot giving the 'truth' (a binary assignment
and/or a continuous score) for each feature, as well as additional
annotations that can be used to stratify the performance calculations.
}
\details{
If adjusted p-values are missing for some methods, for which nominal p-values
are available, the adjusted p-values can be calculated using the
\code{calculate_adjp} function.
}
\section{Slots}{

\describe{
\item{\code{pval}}{A data frame with features as rows and methods as columns,
containing nominal p-values. Missing values (\code{NA}s) are allowed. The
row names should be feature names.}

\item{\code{padj}}{A data frame with features as rows and methods as columns,
containing adjusted p-values. Missing values (\code{NA}s) are allowed. The
row names should be feature names.}

\item{\code{score}}{A data frame with features as rows and methods as columns,
containing generic scores. In case of comparison to a binary truth, larger
values of the scores should correspond to 'more significant' features.
Missing values (\code{NA}s) are allowed. The row names should be feature
names.}

\item{\code{truth}}{A data frame with features as rows columns containing feature
  annotations such as, e.g., binary and continuous truths and additional
  annotations that can be used to stratify the performance calculations. The
  row names should be feature names.}
}}
\author{
Charlotte Soneson
}

