% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAData.R
\name{IBRAData}
\alias{IBRAData}
\title{Constructor function for IBRAData class}
\usage{
IBRAData(pval = data.frame(), padj = data.frame(), score = data.frame(),
  truth = data.frame(), object_to_extend = NULL)
}
\arguments{
\item{pval}{A data frame}

\item{padj}{A data frame}

\item{score}{A data frame}

\item{truth}{A data frame}

\item{object_to_extend}{An IBRAData object}
}
\description{
Constructor function for IBRAData class
}
\author{
Charlotte Soneson
}

