% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAPerformance.R
\docType{class}
\name{IBRAPerformance-class}
\alias{.IBRAPerformance}
\alias{IBRAPerformance-class}
\title{A class to hold performance measures}
\description{
The IBRAPerformance class holds various types of calculated performance
measures. Objects from this class are typically generated by the function
\code{calculate_performance}.
}
\section{Slots}{

\describe{
\item{\code{fdrtpr}}{A data frame containing observed FDR and TPR values at various
adjusted p-value thresholds.}

\item{\code{fdrtprcurve}}{A data frame containing observed FDR and TPR values for a
(potentially large) number of cutoffs applied to a 'score' (that can be
p-value, adjusted p-value or a more general score).}

\item{\code{fdrnbr}}{A data frame containing observed FDR and the number of features
considered to be significant at various adjusted p-value thresholds.}

\item{\code{fdrnbrcurve}}{A data frame containing observed FDR and number of features
considered to be significant for a (potentially large) number of cutoffs
applied to a 'score' (that can be p-value, adjusted p-value or a more
general score).}

\item{\code{tpr}}{A data frame containing observed TPR values at various adjusted
p-value thresholds.}

\item{\code{fpr}}{A data frame containing observed FPR values at various adjusted
p-value thresholds.}

\item{\code{roc}}{A data frame containing observed FPR and TPR values  for a
(potentially large) number of cutoffs applied to a 'score' (that can be
p-value, adjusted p-value or a more general score), which can be used to
generate a ROC curve.}

\item{\code{scatter}}{A data frame containing observed 'scores' (p-values, adjusted
p-values or more general scores) and true scores, which can be used to
generate scatter plots.}

\item{\code{fpc}}{A data frame containing observed numbers of false positive findings
among the N top-ranked features (ranked by p-values, adjusted p-values or
more general scores), for a (potentially large) number of Ns, which can be
used to generate a false positive curve.}

\item{\code{overlap}}{A data frame or list of data frames with binary values
indicating, for each of a number of methods and number of features, whether
the method consider the feature 'positive' (significant, 1) or 'negative'
(non-significant, 0). If it is a list of data frames, each list element
corresponds to one level of a stratifying factor.}

\item{\code{corr}}{A data frame containing observed (Pearson and Spearman)
correlation values between observed and true scores.}

\item{\code{maxsplit}}{A numeric value indicating the largest number of levels to
retain if the results have been stratified by an annotation.}

\item{\code{splv}}{A character string giving the name of the stratification factor,
  "none" if the results are not stratified.}
}}
\author{
Charlotte Soneson
}

