% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAPerformance.R
\name{IBRAPerformance}
\alias{IBRAPerformance}
\title{Constructor function for IBRAPerformance class}
\usage{
IBRAPerformance(fdrtpr = data.frame(), fdrtprcurve = data.frame(),
  fdrnbr = data.frame(), fdrnbrcurve = data.frame(), tpr = data.frame(),
  fpr = data.frame(), splv = "", roc = data.frame(), fpc = data.frame(),
  overlap = data.frame(), maxsplit = 3, corr = data.frame(),
  scatter = data.frame(), object_to_extend = NULL)
}
\arguments{
\item{fdrtpr}{A data frame}

\item{fdrtprcurve}{A data frame}

\item{fdrnbr}{A data frame}

\item{fdrnbrcurve}{A data frame}

\item{tpr}{A data frame}

\item{fpr}{A data frame}

\item{splv}{A character string}

\item{roc}{A data frame}

\item{fpc}{A data frame}

\item{overlap}{A data frame or list}

\item{maxsplit}{A numeric value}

\item{corr}{A data frame}

\item{scatter}{A data frame}

\item{object_to_extend}{An IBRAPerformance object}
}
\description{
Constructor function for IBRAPerformance class
}
\author{
Charlotte Soneson
}

