% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAPlot.R
\docType{class}
\name{IBRAPlot-class}
\alias{.IBRAPlot}
\alias{IBRAPlot-class}
\title{A class to hold performance measures and other attributes, ready for plotting}
\description{
The IBRAPlot class is similar to the IBRAPerformance class in that it holds
various types of calculated performance measures. However, it also contains
other attributes that are necessary for plotting, such as color assignments.
Several IBRAPlot objects can be generated from the same IBRAPerformance
object, without having to go through the potentially time consuming task of
recalculating all performance measures. Objects from this class are typically
generated by the function \code{prepare_data_for_plot}.
}
\section{Slots}{

\describe{
\item{\code{plotcolors}}{A character vector giving the color for each method (or
method-stratification level combination).}

\item{\code{facetted}}{A logical indicating whether the data is prepared for a
  facetted plot (separating different stratification levels into different
  panels) or for displaying all values in one plot panel.}
}}
\author{
Charlotte Soneson
}

