% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{maxsplit}
\alias{maxsplit}
\alias{maxsplit,IBRAPerformance-method}
\alias{maxsplit,IBRAPlot-method}
\alias{maxsplit<-}
\alias{maxsplit<-,IBRAPerformance,numeric-method}
\alias{maxsplit<-,IBRAPlot,numeric-method}
\title{Accessors for the 'maxsplit' slot of an IBRAPerformance object}
\usage{
maxsplit(x, ...)

maxsplit(x, ...) <- value

\S4method{maxsplit}{IBRAPerformance}(x)

\S4method{maxsplit}{IBRAPerformance,numeric}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A numeric value}
}
\description{
Accessors for the 'maxsplit' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

