% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{onlyshared}
\alias{onlyshared}
\alias{onlyshared,IBRAPerformance-method}
\alias{onlyshared,IBRAPlot-method}
\alias{onlyshared<-}
\alias{onlyshared<-,IBRAPerformance,data.frame-method}
\alias{onlyshared<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{onlyshared} slot}
\usage{
onlyshared(x, ...)

onlyshared(x, ...) <- value

\S4method{onlyshared}{IBRAPerformance}(x)

\S4method{onlyshared}{IBRAPerformance,logical}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A logical indicating whether only features that are shared
between result and truth are retained, or if all features in the truth are
used.}
}
\value{
The accessor function returns a logical indicating whether only
  features that are shared between result and truth are retained, or if all
  features in the truth are used.
}
\description{
Accessor and replacement functions for the \code{onlyshared} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr")
head(onlyshared(ibraperf))
}
\author{
Charlotte Soneson
}

