% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{overlap}
\alias{overlap}
\alias{overlap,IBRAPerformance-method}
\alias{overlap,IBRAPlot-method}
\alias{overlap<-}
\alias{overlap<-,IBRAPerformance,list_df-method}
\alias{overlap<-,IBRAPlot,list_df-method}
\title{Accessors for the 'overlap' slot of an IBRAPerformance object}
\usage{
overlap(x, ...)

overlap(x, ...) <- value

\S4method{overlap}{IBRAPerformance}(x)

\S4method{overlap}{IBRAPerformance,list_df}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame or a list}
}
\description{
Accessors for the 'overlap' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

