% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{padj}
\alias{padj}
\alias{padj,IBRAData-method}
\alias{padj<-}
\alias{padj<-,IBRAData,data.frame-method}
\title{Accessors for the 'padj' slot of an IBRAData object}
\usage{
padj(x, ...)

padj(x, ...) <- value

\S4method{padj}{IBRAData}(x)

\S4method{padj}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAData object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'padj' slot of an IBRAData object
}
\author{
Charlotte Soneson
}

