% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_deviation}
\alias{plot_deviation}
\title{Plot deviations}
\usage{
plot_deviation(ibraplot, title = "", stripsize = 15, titlecol = "black",
  plottype = "boxplot", dojitter = TRUE)
}
\arguments{
\item{ibraplot}{An \code{IBRAPlot} object.}

\item{title}{A character string giving the title of the plot.}

\item{stripsize}{A numeric value giving the size of the strip text, when the
results are stratified by an annotation.}

\item{titlecol}{A character string giving the color of the title.}

\item{plottype}{Either "boxplot" or "violin", indicating what type of plot to
make.}

\item{dojitter}{A logical indicating whether to include jittered data points
  or not.}
}
\value{
A ggplot object
}
\description{
Plot the deviations between observed scores and the continuous truth
variable.
}
\examples{
set.seed(123)
score <- data.frame(m1 = runif(100), m2 = runif(100),
                    row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)),
                    row.names = paste0("G", 1:100))
ibradata <- IBRAData(score = score, truth = truth)
ibraperf <- calculate_performance(ibradata, cont_truth = "status",
                                  aspects = "deviation", thrs = c(0.05, 0.1))
ibraplot <- prepare_data_for_plot(ibraperf, keepmethods = c("m1", "m2"),
                                  colorscheme = "Dark2", facetted = FALSE)
plot_deviation(ibraplot)
}
\author{
Charlotte Soneson
}

