% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_fpr}
\alias{plot_fpr}
\title{Plot FPR}
\usage{
plot_fpr(cobraplot, title = "", stripsize = 15, titlecol = "black",
  pointsize = 5, xaxisrange = c(0, 1))
}
\arguments{
\item{cobraplot}{An \code{COBRAPlot} object.}

\item{title}{A character string giving the title of the plot.}

\item{stripsize}{A numeric value giving the size of the strip text, when the
results are stratified by an annotation.}

\item{titlecol}{A character string giving the color of the title.}

\item{pointsize}{A numeric value giving the size of the plot characters.}

\item{xaxisrange}{A numeric vector with two elements, giving the lower and
  upper boundary of the x-axis, respectively.}
}
\value{
A ggplot object
}
\description{
Plot observed false positive rate (FPR) for given adjusted p-value
thresholds.
}
\examples{
data(cobradata_example)
cobraperf <- calculate_performance(cobradata_example, binary_truth = "status",
                                  aspects = "fpr")
cobraplot <- prepare_data_for_plot(cobraperf, colorscheme = "Dark2",
                                  incltruth = TRUE)
plot_fpr(cobraplot, xaxisrange = c(0, 0.25))
}
\author{
Charlotte Soneson
}

