% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot Venn diagram}
\usage{
plot_overlap(ibraplot, ...)
}
\arguments{
\item{ibraplot}{An \code{IBRAPlot} object.}

\item{...}{Additional arguments to \code{limma::vennDiagram}.}
}
\value{
Nothing, displays a graph
}
\description{
Plot a Venn diagram showing the overlaps among sets of significant feature
for a given adjusted p-value threshold. Optionally, the truth can be included
as a "perfect" method. Note that maximally five methods (including the truth,
if applicable) can be compared.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "overlap")
ibraplot <- prepare_data_for_plot(ibraperf, colorscheme = "Dark2",
                                  incltruth = TRUE)
plot_overlap(ibraplot)
}
\author{
Charlotte Soneson
}

