% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot scatter plots}
\usage{
plot_scatter(ibraplot, title = "", stripsize = 10, titlecol = "black",
  pointsize = 3, doflip = FALSE)
}
\arguments{
\item{ibraplot}{An \code{IBRAPlot} object.}

\item{title}{A character string giving the title of the plot.}

\item{stripsize}{A numeric value giving the size of the strip text, when the
results are stratified by an annotation.}

\item{titlecol}{A character string giving the color of the title.}

\item{pointsize}{A numeric value giving the size of the plot characters.}

\item{doflip}{A logical indicating whether to flip the axes when results are
  stratified by an annotation. By default (\code{doflip = FALSE}),
  stratification levels are shown as columns and methods as rows in the plot.}
}
\value{
A ggplot object
}
\description{
Plot scatter plots, indicating the relationship between observed values and a
continuous truth.
}
\examples{
set.seed(123)
padj <- data.frame(m1 = runif(100), m2 = runif(100), row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)), row.names = paste0("G", 1:100))
ibradata <- IBRAData(padj = padj, truth = truth)
ibraperf <- calculate_performance(ibradata, cont_truth = "status",
                                  aspects = "scatter")
ibraplot <- prepare_data_for_plot(ibraperf, keepmethods = c("m1", "m2"),
                                  colorscheme = "Dark2", facetted = TRUE)
plot_scatter(ibraplot, doflip = TRUE)
}
\author{
Charlotte Soneson
}

