% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_tpr}
\alias{plot_tpr}
\title{Plot TPR}
\usage{
plot_tpr(ibraplot, title = "", stripsize = 15, titlecol = "black",
  pointsize = 5, xaxisrange = c(0, 1))
}
\arguments{
\item{ibraplot}{An \code{IBRAPlot} object.}

\item{title}{A character string giving the title of the plot.}

\item{stripsize}{A numeric value giving the size of the strip text, when the
results are stratified by an annotation.}

\item{titlecol}{A character string giving the color of the title.}

\item{pointsize}{A numeric value giving the size of the plot characters.}

\item{xaxisrange}{A numeric vector with two elements, giving the lower and
  upper boundary of the x-axis, respectively.}
}
\value{
A ggplot object
}
\description{
Plot observed true positive rate (TPR) for given adjusted p-value thresholds.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "tpr")
ibraplot <- prepare_data_for_plot(ibraperf, colorscheme = "Dark2",
                                  incltruth = TRUE)
plot_tpr(ibraplot)
}
\author{
Charlotte Soneson
}

