% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{pval}
\alias{pval}
\alias{pval,IBRAData-method}
\alias{pval<-}
\alias{pval<-,IBRAData,data.frame-method}
\title{Accessors for the 'pval' slot of an IBRAData object}
\usage{
pval(x, ...)

pval(x, ...) <- value

\S4method{pval}{IBRAData}(x)

\S4method{pval}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAData object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'pval' slot of an IBRAData object
}
\author{
Charlotte Soneson
}

