% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{roc}
\alias{roc}
\alias{roc,IBRAPerformance-method}
\alias{roc,IBRAPlot-method}
\alias{roc<-}
\alias{roc<-,IBRAPerformance,data.frame-method}
\alias{roc<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'roc' slot of an IBRAPerformance object}
\usage{
roc(x, ...)

roc(x, ...) <- value

\S4method{roc}{IBRAPerformance}(x)

\S4method{roc}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'roc' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

