% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{scatter}
\alias{scatter}
\alias{scatter,IBRAPerformance-method}
\alias{scatter,IBRAPlot-method}
\alias{scatter<-}
\alias{scatter<-,IBRAPerformance,data.frame-method}
\alias{scatter<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{scatter} slot}
\usage{
scatter(x, ...)

scatter(x, ...) <- value

\S4method{scatter}{IBRAPerformance}(x)

\S4method{scatter}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information necessary to generate scatter
plots of observed vs true values for each method and each stratification
level.}
}
\value{
The accessor function returns a data frame giving information
  necessary to generate scatter plots of observed vs true values for each
  method and each stratification level.
}
\description{
Accessor and replacement functions for the \code{scatter} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, cont_truth = "logFC",
                                  aspects = "scatter")
head(scatter(ibraperf))
}
\author{
Charlotte Soneson
}

