% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{score}
\alias{score}
\alias{score,IBRAData-method}
\alias{score<-}
\alias{score<-,IBRAData,data.frame-method}
\title{Accessor and replacement functions for \code{score} slot}
\usage{
score(x, ...)

score(x, ...) <- value

\S4method{score}{IBRAData}(x)

\S4method{score}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing scores for each feature and each method.}
}
\description{
Accessor and replacement functions for the \code{score} slot in an
\code{IBRAData} object.
}
\author{
Charlotte Soneson
}

