% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{splv}
\alias{splv}
\alias{splv,IBRAPerformance-method}
\alias{splv,IBRAPlot-method}
\alias{splv<-}
\alias{splv<-,IBRAPerformance,character-method}
\alias{splv<-,IBRAPlot,character-method}
\title{Accessors for the 'splv' slot of an IBRAPerformance object}
\usage{
splv(x, ...)

splv(x, ...) <- value

\S4method{splv}{IBRAPerformance}(x)

\S4method{splv}{IBRAPerformance,character}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A character string}
}
\description{
Accessors for the 'splv' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

