% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{truth}
\alias{truth}
\alias{truth,IBRAData-method}
\alias{truth<-}
\alias{truth<-,IBRAData,data.frame-method}
\title{Accessors for the 'truth' slot of an IBRAData object}
\usage{
truth(x, ...)

truth(x, ...) <- value

\S4method{truth}{IBRAData}(x)

\S4method{truth}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAData object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'truth' slot of an IBRAData object
}
\author{
Charlotte Soneson
}

