% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{basemethods}
\alias{basemethods}
\alias{basemethods,IBRAPerformance-method}
\alias{basemethods,IBRAPlot-method}
\title{Accessors for the basemethods contained in an IBRAPerformance object}
\usage{
basemethods(x, ...)

\S4method{basemethods}{IBRAPerformance}(x)
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}
}
\description{
Accessors for the basemethods contained in an IBRAPerformance object
}
\author{
Charlotte Soneson
}

