% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{calculate_adjp}
\alias{calculate_adjp}
\title{Calculate adjusted p-values}
\usage{
calculate_adjp(ibradata, method = "BH")
}
\arguments{
\item{ibradata}{An \code{IBRAData} object.}

\item{method}{A character string giving the method (selected from
  \code{p.adjust.methods()}) that will be used to perform the adjustment.}
}
\value{
An \code{IBRAData} object, extended with the calculated adjusted p-values.
}
\description{
Calculate adjusted p-values for methods where only nominal p-values are
available in an \code{IBRAData} object.
}
\examples{
set.seed(123)
pval <- data.frame(m1 = runif(100), m2 = runif(100), row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)), row.names = paste0("G", 1:100))
ibradata <- IBRAData(pval = pval, truth = truth)
ibradata <- calculate_adjp(ibradata, method = "BH")
}
\author{
Charlotte Soneson
}

