% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{calculate_adjp}
\alias{calculate_adjp}
\title{Calculate adjusted p-values}
\usage{
calculate_adjp(ibradata, method = "BH")
}
\arguments{
\item{ibradata}{An \code{IBRAData} object.}

\item{method}{A character string giving the method (selected from
  \code{p.adjust.methods()}) that will be used to perform the adjustment.}
}
\value{
An \code{IBRAData} object, extended with the calculated adjusted
  p-values.
}
\description{
Calculate adjusted p-values for methods where only nominal p-values are
available in an \code{IBRAData} object.
}
\examples{
data(ibradata_example)
ibradata_example <- calculate_adjp(ibradata_example, method = "BH")
}
\author{
Charlotte Soneson
}

