% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAPlot.R
\docType{methods}
\name{coerce}
\alias{IBRAPlot,IBRAPerformance-method}
\alias{coerce}
\alias{coerce,}
\alias{coerce,IBRAPerformance,IBRAPlot-method}
\alias{coerce,IBRAPlot,IBRAPerformance-method}
\title{Convert an object to another class}
\arguments{
\item{from}{The object that is to be coerced into another class.}
}
\description{
Convert object between \code{IBRAPerformance} and \code{IBRAPlot} classes.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr")
ibraplot <- prepare_data_for_plot(ibraperf)

## Coerce IBRAPerformance object into IBRAPlot object
as(ibraperf, "IBRAPlot")

## Coerce IBRAPlot object into IBRAPerformance object
as(ibraplot, "IBRAPerformance")
}
\author{
Charlotte Soneson
}

