% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAPlot.R
\docType{methods}
\name{coerce}
\alias{IBRAPlot,IBRAPerformance-method}
\alias{coerce}
\alias{coerce,}
\alias{coerce,IBRAPerformance,IBRAPlot-method}
\alias{coerce,IBRAPlot,IBRAPerformance-method}
\title{Convert an object to another class}
\arguments{
\item{from}{The object that is to be coerced into another class.}
}
\description{
Convert object between \code{IBRAPerformance} and \code{IBRAPlot} classes.
}
\examples{
set.seed(123)
padj <- data.frame(m1 = runif(100), m2 = runif(100), row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)), row.names = paste0("G", 1:100))
ibradata <- IBRAData(padj = padj, truth = truth)
ibraperf <- calculate_performance(ibradata, binary_truth = "status",
                                  aspects = c("fdrtpr", "fdrtprcurve",
                                              "tpr", "roc"),
                                  thrs = c(0.01, 0.05, 0.1), splv = "none")

## Coerce IBRAPerformance object into IBRAPlot object
ibraplot <- as(ibraperf, "IBRAPlot")

## Coerce IBRAPlot object into IBRAPerformance object
ibraperf2 <- as(ibraplot, "IBRAPerformance")
}
\author{
Charlotte Soneson
}

