% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{corr}
\alias{corr}
\alias{corr,IBRAPerformance-method}
\alias{corr<-}
\alias{corr<-,IBRAPerformance,data.frame-method}
\title{Accessor and replacement functions for \code{corr} slot}
\usage{
corr(x, ...)

corr(x, ...) <- value

\S4method{corr}{IBRAPerformance}(x)

\S4method{corr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving correlation values for each method and each
stratification level.}
}
\description{
Accessor and replacement functions for the \code{corr} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\author{
Charlotte Soneson
}

