% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAData.R, R/IBRAPerformance.R, R/IBRAPlot.R
\docType{methods}
\name{Extract}
\alias{Extract}
\alias{[}
\alias{[,IBRAData-method}
\alias{[,IBRAPerformance-method}
\alias{[,IBRAPlot-method}
\alias{\S4method{[}{IBRAData,ANY,ANY}}
\alias{\S4method{[}{IBRAPerformance,ANY,ANY}}
\alias{\S4method{[}{IBRAPlot,ANY,ANY}}
\title{Subsetting an IBRAData, IBRAPerformance or IBRAPlot object}
\usage{
\S4method{[}{IBRAData,ANY,ANY}(x, i, j = "missing", drop = "missing")

\S4method{[}{IBRAPerformance,ANY,ANY}(x, i = "missing", j, drop = "missing")

\S4method{[}{IBRAPlot,ANY,ANY}(x, i = "missing", j, drop = "missing")
}
\arguments{
\item{x}{An IBRAData, IBRAPerformance or IBRAPlot object}

\item{i}{For IBRAData objects, a character vector of feature names to retain}

\item{j}{For IBRAPerformance and IBRAPlot objects, a character vector with
method names to retain}

\item{drop}{not used}
}
\description{
Subsetting an IBRAData, IBRAPerformance or IBRAPlot object
}

