% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAData.R, R/IBRAPerformance.R, R/IBRAPlot.R
\docType{methods}
\name{Extract}
\alias{Extract}
\alias{[}
\alias{[,IBRAData-method}
\alias{[,IBRAPerformance-method}
\alias{[,IBRAPlot-method}
\alias{\S4method{[}{IBRAData,ANY,ANY}}
\alias{\S4method{[}{IBRAPerformance,ANY,ANY}}
\alias{\S4method{[}{IBRAPlot,ANY,ANY}}
\title{Subsetting \code{IBRAData}, \code{IBRAPerformance} or \code{IBRAPlot} objects}
\usage{
\S4method{[}{IBRAData}(x, i, j = "missing", drop = "missing")

\S4method{[}{IBRAPerformance}(x, i = "missing", j, drop = "missing")

\S4method{[}{IBRAPlot}(x, i = "missing", j, drop = "missing")
}
\arguments{
\item{x}{An \code{IBRAData}, \code{IBRAPerformance} or \code{IBRAPlot}
object.}

\item{i}{For \code{IBRAData} objects, a character vector of feature names to
retain.}

\item{j}{For \code{IBRAPerformance} and \code{IBRAPlot} objects, a character
vector with method names to retain.}

\item{drop}{not used.}
}
\value{
A subset of the original object, of the same class
}
\description{
Functions to subset \code{IBRAData}, \code{IBRAPerformance} or
\code{IBRAPlot} objects. \code{IBRAData} objects are subset by features
(rows), while \code{IBRAPerformance} and \code{IBRAPlot} objects are subset
by methods (columns). Numeric indices are not allowed, since not all slots
may be arranged in the same order.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr")
ibraperf[, c("voom")]
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr")
ibraplot <- prepare_data_for_plot(ibraperf)
ibraplot[, c("voom")]
}

