% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPlot.R
\docType{methods}
\name{facetted}
\alias{facetted}
\alias{facetted,IBRAPlot-method}
\alias{facetted<-}
\alias{facetted<-,IBRAPlot,logical-method}
\title{Accessors for the 'facetted' slot of an IBRAPlot object}
\usage{
facetted(x, ...)

facetted(x, ...) <- value

\S4method{facetted}{IBRAPlot}(x)

\S4method{facetted}{IBRAPlot,logical}(x) <- value
}
\arguments{
\item{x}{An IBRAPlot object}

\item{...}{Additional arguments}

\item{value}{A logical}
}
\description{
Accessors for the 'facetted' slot of an IBRAPlot object
}
\author{
Charlotte Soneson
}

