% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPlot.R
\docType{methods}
\name{facetted}
\alias{facetted}
\alias{facetted,IBRAPlot-method}
\alias{facetted<-}
\alias{facetted<-,IBRAPlot,logical-method}
\title{Accessor and replacement functions for \code{facetted} slot}
\usage{
facetted(x, ...)

facetted(x, ...) <- value

\S4method{facetted}{IBRAPlot}(x)

\S4method{facetted}{IBRAPlot,logical}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A logical value, indicating whether the object is formatted for
facetted plots (visualizing each stratification level in a separate panel)
or not.}
}
\value{
The accessor function returns a logical value, indicating whether the
  object is formatted for facetted plots (visualizing each stratification
  level in a separate panel) or not.
}
\description{
Accessor and replacement functions for the \code{facetted} slot in an
\code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr")
ibraplot <- prepare_data_for_plot(ibraperf)
facetted(ibraplot)
}
\author{
Charlotte Soneson
}

