% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrnbr}
\alias{fdrnbr}
\alias{fdrnbr,IBRAPerformance-method}
\alias{fdrnbr,IBRAPlot-method}
\alias{fdrnbr<-}
\alias{fdrnbr<-,IBRAPerformance,data.frame-method}
\alias{fdrnbr<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'fdrnbr' slot of an IBRAPerformance object}
\usage{
fdrnbr(x, ...)

fdrnbr(x, ...) <- value

\S4method{fdrnbr}{IBRAPerformance}(x)

\S4method{fdrnbr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'fdrnbr' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

