% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrnbrcurve}
\alias{fdrnbrcurve}
\alias{fdrnbrcurve,IBRAPerformance-method}
\alias{fdrnbrcurve,IBRAPlot-method}
\alias{fdrnbrcurve<-}
\alias{fdrnbrcurve<-,IBRAPerformance,data.frame-method}
\alias{fdrnbrcurve<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'fdrnbrcurve' slot of an IBRAPerformance object}
\usage{
fdrnbrcurve(x, ...)

fdrnbrcurve(x, ...) <- value

\S4method{fdrnbrcurve}{IBRAPerformance}(x)

\S4method{fdrnbrcurve}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'fdrnbrcurve' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

