% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrnbrcurve}
\alias{fdrnbrcurve}
\alias{fdrnbrcurve,IBRAPerformance-method}
\alias{fdrnbrcurve,IBRAPlot-method}
\alias{fdrnbrcurve<-}
\alias{fdrnbrcurve<-,IBRAPerformance,data.frame-method}
\alias{fdrnbrcurve<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{fdrnbrcurve} slot}
\usage{
fdrnbrcurve(x, ...)

fdrnbrcurve(x, ...) <- value

\S4method{fdrnbrcurve}{IBRAPerformance}(x)

\S4method{fdrnbrcurve}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information necessary to generate curves of
observed FDR vs number of features called positive for each method and each
stratification level.}
}
\description{
Accessor and replacement functions for the \code{fdrnbrcurve} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\author{
Charlotte Soneson
}

