% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrtpr}
\alias{fdrtpr}
\alias{fdrtpr,IBRAPerformance-method}
\alias{fdrtpr,IBRAPlot-method}
\alias{fdrtpr<-}
\alias{fdrtpr<-,IBRAPerformance,data.frame-method}
\alias{fdrtpr<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{fdrtpr} slot}
\usage{
fdrtpr(x, ...)

fdrtpr(x, ...) <- value

\S4method{fdrtpr}{IBRAPerformance}(x)

\S4method{fdrtpr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information about the observed FPR and TPR
for each method and each stratification level, at various adjusted p-value
thresholds.}
}
\value{
The accessor function returns a data frame giving information about
  the observed FPR and TPR for each method and each stratification level, at
  various adjusted p-value thresholds.
}
\description{
Accessor and replacement functions for the \code{fdrtpr} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr")
head(fdrtpr(ibraperf))
}
\author{
Charlotte Soneson
}

