% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fdrtprcurve}
\alias{fdrtprcurve}
\alias{fdrtprcurve,IBRAPerformance-method}
\alias{fdrtprcurve,IBRAPlot-method}
\alias{fdrtprcurve<-}
\alias{fdrtprcurve<-,IBRAPerformance,data.frame-method}
\alias{fdrtprcurve<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'fdrtprcurve' slot of an IBRAPerformance object}
\usage{
fdrtprcurve(x, ...)

fdrtprcurve(x, ...) <- value

\S4method{fdrtprcurve}{IBRAPerformance}(x)

\S4method{fdrtprcurve}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'fdrtprcurve' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

