% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fpc}
\alias{fpc}
\alias{fpc,IBRAPerformance-method}
\alias{fpc,IBRAPlot-method}
\alias{fpc<-}
\alias{fpc<-,IBRAPerformance,data.frame-method}
\alias{fpc<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'fpc' slot of an IBRAPerformance object}
\usage{
fpc(x, ...)

fpc(x, ...) <- value

\S4method{fpc}{IBRAPerformance}(x)

\S4method{fpc}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'fpc' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

