% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fpc}
\alias{fpc}
\alias{fpc,IBRAPerformance-method}
\alias{fpc,IBRAPlot-method}
\alias{fpc<-}
\alias{fpc<-,IBRAPerformance,data.frame-method}
\alias{fpc<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{fpc} slot}
\usage{
fpc(x, ...)

fpc(x, ...) <- value

\S4method{fpc}{IBRAPerformance}(x)

\S4method{fpc}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information necessary to generate false
positive curves for each method and each stratification level.}
}
\description{
Accessor and replacement functions for the \code{fpc} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\author{
Charlotte Soneson
}

