% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{fpr}
\alias{fpr}
\alias{fpr,IBRAPerformance-method}
\alias{fpr,IBRAPlot-method}
\alias{fpr<-}
\alias{fpr<-,IBRAPerformance,data.frame-method}
\alias{fpr<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'fpr' slot of an IBRAPerformance object}
\usage{
fpr(x, ...)

fpr(x, ...) <- value

\S4method{fpr}{IBRAPerformance}(x)

\S4method{fpr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'fpr' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

