% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{IBRAapp}
\alias{IBRAapp}
\title{Run interactive shiny app to visualize results}
\usage{
IBRAapp(ibradata = NULL)
}
\arguments{
\item{ibradata}{An (optional) IBRAData object. If not given, the user can
load results from text files.}
}
\description{
Run interactive shiny app to visualize results
}
\author{
Charlotte Soneson
}

