% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAData.R
\docType{class}
\name{IBRAData-class}
\alias{.IBRAData}
\alias{IBRAData}
\alias{IBRAData-class}
\alias{IBRAData_from_simresults}
\alias{IBRAData_from_text}
\alias{IBRAData_to_simresults}
\alias{IBRAData_to_text}
\title{\code{IBRAData} object and constructor}
\usage{
IBRAData(pval = data.frame(), padj = data.frame(), score = data.frame(),
  truth = data.frame(), object_to_extend = NULL)

IBRAData_from_text(truth_file, result_files, feature_id)

IBRAData_to_text(ibradata, truth_file, result_files, feature_id)

IBRAData_from_simresults(simres)

IBRAData_to_simresults(ibradata, binary_truth, strat = NULL)
}
\arguments{
\item{pval}{A data frame with features as rows and methods as columns,
containing nominal p-values. Missing values (\code{NA}s) are allowed. The
row names should be feature names.}

\item{padj}{A data frame with features as rows and methods as columns,
containing adjusted p-values. Missing values (\code{NA}s) are allowed. The
row names should be feature names.}

\item{score}{A data frame with features as rows and methods as columns,
containing generic scores. In case of comparison to a binary truth, larger
values of the scores should correspond to 'more significant' features.
Missing values (\code{NA}s) are allowed. The row names should be feature
names.}

\item{truth}{A data frame with features as rows columns containing feature
annotations such as, e.g., binary and continuous truths and additional
annotations that can be used to stratify the performance calculations. The
row names should be feature names.}

\item{object_to_extend}{An \code{IBRAData} object to extend with the provided
  information.}

\item{truth_file}{A character string giving the path to a file with true
labels and other feature annotations.}

\item{result_files}{A character vector giving path(s) to file(s) with results
(p-values, adjusted p-values, scores) for one or more methods. The column
names of these files must be of the form "method:measure", where measure is
one of P, adjP or score, depending on what is given in the column.}

\item{feature_id}{A character string giving the name of the column in the
truth and result files that encodes the feature identifier.}

\item{ibradata}{An \code{IBRAData} object}

\item{simres}{A \code{benchmarkR::SimResults} object}

\item{binary_truth}{The name of the column in truth(ibradata) that contains
the binary truth variable.}

\item{strat}{(optional) The name of the column in truth(ibradata) that
contains the stratification variable.}
}
\description{
The \code{IBRAData} class contains slots to hold calculated p-values,
adjusted p-values and general 'scores' for a set of features. The slots can
contain values from multiple methods, and each method can contribute to one
or more slots. The class also contains a slot giving the 'truth' (a binary
assignment and/or a continuous score) for each feature, as well as additional
annotations that can be used to stratify the performance calculations.
}
\details{
If adjusted p-values are missing for some methods, for which nominal p-values
are available, the adjusted p-values can be calculated using the
\code{\link{calculate_adjp}} function.

The text files generated by \code{IBRAData_to_text} can be used as input to
\code{IBRAapp}, when it is called without an input argument.
}
\author{
Charlotte Soneson
}

