% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IBRAPlot.R
\name{IBRAPlot}
\alias{IBRAPlot}
\title{Constructor function for IBRAPlot class}
\usage{
IBRAPlot(fdrtpr = data.frame(), fdrtprcurve = data.frame(),
  fdrnbr = data.frame(), corr = data.frame(), fdrnbrcurve = data.frame(),
  tpr = data.frame(), fpr = data.frame(), roc = data.frame(),
  scatter = data.frame(), fpc = data.frame(), overlap = data.frame(),
  plotcolors = c(), splv = "", maxsplit = NA_integer_, facetted = TRUE,
  object_to_extend = NULL)
}
\arguments{
\item{fdrtpr}{A data frame}

\item{fdrtprcurve}{A data frame}

\item{fdrnbr}{A data frame}

\item{corr}{A data frame}

\item{fdrnbrcurve}{A data frame}

\item{tpr}{A data frame}

\item{fpr}{A data frame}

\item{roc}{A data frame}

\item{scatter}{A data frame}

\item{fpc}{A data frame}

\item{overlap}{A data frame}

\item{plotcolors}{A character vector}

\item{splv}{A character string}

\item{maxsplit}{A numeric value}

\item{facetted}{A logical value}

\item{object_to_extend}{An IBRAPlot object}
}
\description{
Constructor function for IBRAPlot class
}
\author{
Charlotte Soneson
}

