% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{maxsplit}
\alias{maxsplit}
\alias{maxsplit,IBRAPerformance-method}
\alias{maxsplit,IBRAPlot-method}
\alias{maxsplit<-}
\alias{maxsplit<-,IBRAPerformance,numeric-method}
\alias{maxsplit<-,IBRAPlot,numeric-method}
\title{Accessor and replacement functions for \code{maxsplit} slot}
\usage{
maxsplit(x, ...)

maxsplit(x, ...) <- value

\S4method{maxsplit}{IBRAPerformance}(x)

\S4method{maxsplit}{IBRAPerformance,numeric}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A numeric value giving the maximal number of strata to retain.}
}
\value{
The accessor function returns a numeric value giving the maximal
  number of strata to retain.
}
\description{
Accessor and replacement functions for the \code{maxsplit} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr", splv = "expr_cat",
                                  maxsplit = 3)
maxsplit(ibraperf)
}
\author{
Charlotte Soneson
}

