% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{overlap}
\alias{overlap}
\alias{overlap,IBRAPerformance-method}
\alias{overlap,IBRAPlot-method}
\alias{overlap<-}
\alias{overlap<-,IBRAPerformance,list_df-method}
\alias{overlap<-,IBRAPlot,list_df-method}
\title{Accessor and replacement functions for \code{overlap} slot}
\usage{
overlap(x, ...)

overlap(x, ...) <- value

\S4method{overlap}{IBRAPerformance}(x)

\S4method{overlap}{IBRAPerformance,list_df}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame or a list, giving information about which feature
that are classified as 'positive' by each method and for each
stratification level.}
}
\value{
The accessor function returns a data frame or a list, giving
  information about which feature that are classified as 'positive' by each
  method and for each stratification level.
}
\description{
Accessor and replacement functions for the \code{overlap} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "overlap")
head(overlap(ibraperf))
}
\author{
Charlotte Soneson
}

