% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/COBRAData.R
\docType{methods}
\name{padj}
\alias{padj}
\alias{padj<-}
\alias{padj}
\alias{padj,COBRAData-method}
\alias{padj<-,COBRAData,data.frame-method}
\alias{padj}
\title{Accessor and replacement functions for \code{padj} slot}
\usage{
padj(x, ...)

padj(x, ...) <- value

\S4method{padj}{COBRAData}(x)

\S4method{padj}{COBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{A \code{COBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing adjusted p-values for each feature and
each method.}
}
\value{
The accessor function returns a data frame containing adjusted
  p-values for each feature and each method.
}
\description{
Accessor and replacement functions for the \code{padj} slot in a
\code{COBRAData} object.
}
\examples{
data(cobradata_example)
head(padj(cobradata_example))
}
\author{
Charlotte Soneson
}
