% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{padj}
\alias{padj}
\alias{padj,IBRAData-method}
\alias{padj<-}
\alias{padj<-,IBRAData,data.frame-method}
\title{Accessor and replacement functions for \code{padj} slot}
\usage{
padj(x, ...)

padj(x, ...) <- value

\S4method{padj}{IBRAData}(x)

\S4method{padj}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing adjusted p-values for each feature and
each method.}
}
\description{
Accessor and replacement functions for the \code{padj} slot in an
\code{IBRAData} object.
}
\author{
Charlotte Soneson
}

