% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_fdrtprcurve}
\alias{plot_fdrtprcurve}
\title{Plot TPR vs FDR}
\usage{
plot_fdrtprcurve(ibraplot, title = "", stripsize = 15, titlecol = "black",
  pointsize = 5, xaxisrange = c(0, 1), yaxisrange = c(0, 1),
  plottype = c("curve", "points"))
}
\arguments{
\item{ibraplot}{An \code{IBRAPlot} object.}

\item{title}{A character string giving the title of the plot.}

\item{stripsize}{A numeric value giving the size of the strip text, when the
results are stratified by an annotation.}

\item{titlecol}{A character string giving the color of the title.}

\item{pointsize}{A numeric value giving the size of the plot characters.}

\item{xaxisrange}{A numeric vector with two elements, giving the lower and
upper boundary of the x-axis, respectively.}

\item{yaxisrange}{A numeric vector with two elements, giving the lower and
upper boundary of the y-axis, respectively.}

\item{plottype}{A character vector giving the type of plot to construct. Can
  be any combination of the two elements "curve" and "points".}
}
\value{
A ggplot object
}
\description{
Plot observed true positive rate (TPR) vs observed false discovery rate
(FDR), for given adjusted p-value thresholds and/or as curves traced out by
considering all threshold values.
}
\examples{
set.seed(123)
padj <- data.frame(m1 = runif(100), m2 = runif(100),
                   row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)), row.names = paste0("G", 1:100))
ibradata <- IBRAData(padj = padj, truth = truth)
ibraperf <- calculate_performance(ibradata, binary_truth = "status",
                                  aspects = c("fdrtpr", "fdrtprcurve"),
                                  thrs = c(0.05, 0.1))
ibraplot <- prepare_data_for_plot(ibraperf, keepmethods = c("m1", "m2"),
                                  colorscheme = "Dark2", facetted = FALSE)
plot_fdrtprcurve(ibraplot, plottype = c("curve", "points"))
}
\author{
Charlotte Soneson
}

