% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot Venn diagram}
\usage{
plot_overlap(ibraplot, ...)
}
\arguments{
\item{ibraplot}{An object of class IBRAPlot}

\item{...}{Additional arguments to limma::vennDiagram}
}
\value{
Nothing, displays a graph
}
\description{
Plot a Venn diagram showing the overlaps among sets of significant feature
for a given adjusted p-value threshold. Optionally, the truth can be included
as a "perfect" method. Note that maximally five methods (including the truth,
if applicable) can be compared.
}
\examples{
set.seed(123)
padj <- data.frame(m1 = runif(100), m2 = runif(100), row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)), row.names = paste0("G", 1:100))
ibradata <- IBRAData(padj = padj, truth = truth)
ibraperf <- calculate_performance(ibradata, binary_truth = "status",
                                  aspects = "overlap", thr_venn = 0.1)
ibraplot <- prepare_data_for_plot(ibraperf, keepmethods = c("m1", "m2"),
                                  colorscheme = "Dark2", incltruth = TRUE)
plot_overlap(ibraplot)
}
\author{
Charlotte Soneson
}

