% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curves}
\usage{
plot_roc(ibraplot, title = "", stripsize = 15, titlecol = "black",
  xaxisrange = c(0, 1), yaxisrange = c(0, 1))
}
\arguments{
\item{ibraplot}{An \code{IBRAPlot} object.}

\item{title}{A character string giving the title of the plot.}

\item{stripsize}{A numeric value giving the size of the strip text, when the
results are stratified by an annotation.}

\item{titlecol}{A character string giving the color of the title.}

\item{xaxisrange}{A numeric vector with two elements, giving the lower and
upper boundary of the x-axis, respectively.}

\item{yaxisrange}{A numeric vector with two elements, giving the lower and
  upper boundary of the y-axis, respectively.}
}
\value{
A ggplot object
}
\description{
Plot receiver operating characteristics (ROC) curves.
}
\examples{
set.seed(123)
padj <- data.frame(m1 = runif(100), m2 = runif(100), row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)), row.names = paste0("G", 1:100))
ibradata <- IBRAData(padj = padj, truth = truth)
ibraperf <- calculate_performance(ibradata, binary_truth = "status",
                                  aspects = "roc")
ibraplot <- prepare_data_for_plot(ibraperf, keepmethods = c("m1", "m2"),
                                  colorscheme = "Dark2", facetted = FALSE)
plot_roc(ibraplot, xaxisrange = c(0, 1), yaxisrange = c(0, 1))
}
\author{
Charlotte Soneson
}

