% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_tpr}
\alias{plot_tpr}
\title{Plot TPR}
\usage{
plot_tpr(ibraplot, title = "", stripsize = 15, titlecol = "black",
  pointsize = 5, xaxisrange = c(0, 1))
}
\arguments{
\item{ibraplot}{An object of class IBRAPlot}

\item{title}{A character string giving the title of the plot}

\item{stripsize}{A numeric value giving the size of the strip text, when the
results are stratified by an annotation}

\item{titlecol}{A character string giving the color of the title}

\item{pointsize}{A numeric value giving the size of the plot characters}

\item{xaxisrange}{A numeric vector with two elements, giving the lower and
  upper boundary of the x-axis, respectively}
}
\value{
A ggplot object
}
\description{
Plot observed true positive rate (TPR) for given adjusted p-value thresholds.
}
\examples{
set.seed(123)
padj <- data.frame(m1 = runif(100), m2 = runif(100), row.names = paste0("G", 1:100))
truth <- data.frame(status = round(runif(100)), row.names = paste0("G", 1:100))
ibradata <- IBRAData(padj = padj, truth = truth)
ibraperf <- calculate_performance(ibradata, binary_truth = "status",
                                  aspects = "tpr", thrs = c(0.05, 0.1))
ibraplot <- prepare_data_for_plot(ibraperf, keepmethods = c("m1", "m2"),
                                  colorscheme = "Dark2", facetted = FALSE)
plot_tpr(ibraplot, xaxisrange = c(0, 0.5))
}
\author{
Charlotte Soneson
}

