% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPlot.R
\docType{methods}
\name{plotcolors}
\alias{plotcolors}
\alias{plotcolors,IBRAPlot-method}
\alias{plotcolors<-}
\alias{plotcolors<-,IBRAPlot,character-method}
\title{Accessor and replacement functions for \code{plotcolors} slot}
\usage{
plotcolors(x, ...)

plotcolors(x, ...) <- value

\S4method{plotcolors}{IBRAPlot}(x)

\S4method{plotcolors}{IBRAPlot,character}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A character vector giving the colors assigned to each of the
methods (or method/stratification level combinations) represented in the
\code{IBRAPlot} object.}
}
\value{
The accessor function returns a character vector giving the colors
  assigned to each of the methods (or method/stratification level
  combinations) represented in the \code{IBRAPlot} object.
}
\description{
Accessor and replacement functions for the \code{plotcolors} slot in an
\code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr")
ibraplot <- prepare_data_for_plot(ibraperf)
plotcolors(ibraplot)
}
\author{
Charlotte Soneson
}

