% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPlot.R
\docType{methods}
\name{plotcolors}
\alias{plotcolors}
\alias{plotcolors,IBRAPlot-method}
\alias{plotcolors<-}
\alias{plotcolors<-,IBRAPlot,character-method}
\title{Accessors for the 'plotcolors' slot of an IBRAPlot object}
\usage{
plotcolors(x, ...)

plotcolors(x, ...) <- value

\S4method{plotcolors}{IBRAPlot}(x)

\S4method{plotcolors}{IBRAPlot,character}(x) <- value
}
\arguments{
\item{x}{An IBRAPlot object}

\item{...}{Additional arguments}

\item{value}{A character vector}
}
\description{
Accessors for the 'plotcolors' slot of an IBRAPlot object
}
\author{
Charlotte Soneson
}

