% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{pval}
\alias{pval}
\alias{pval,IBRAData-method}
\alias{pval<-}
\alias{pval<-,IBRAData,data.frame-method}
\title{Accessor and replacement functions for \code{pval} slot}
\usage{
pval(x, ...)

pval(x, ...) <- value

\S4method{pval}{IBRAData}(x)

\S4method{pval}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing p-values for each feature and each
method.}
}
\description{
Accessor and replacement functions for the \code{pval} slot in an
\code{IBRAData} object.
}
\author{
Charlotte Soneson
}

