% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/COBRAData.R
\docType{methods}
\name{pval}
\alias{pval}
\alias{pval<-}
\alias{pval,COBRAData-method}
\alias{pval<-,COBRAData,data.frame-method}
\title{Accessor and replacement functions for \code{pval} slot}
\usage{
pval(x, ...)

pval(x, ...) <- value

\S4method{pval}{COBRAData}(x)

\S4method{pval}{COBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{A \code{COBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing p-values for each feature and each
method.}
}
\value{
The accessor function returns a data frame containing p-values for
  each feature and each method.
}
\description{
Accessor and replacement functions for the \code{pval} slot in a
\code{COBRAData} object.
}
\examples{
data(cobradata_example)
head(pval(cobradata_example))
}
\author{
Charlotte Soneson
}
