% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{roc}
\alias{roc}
\alias{roc,IBRAPerformance-method}
\alias{roc,IBRAPlot-method}
\alias{roc<-}
\alias{roc<-,IBRAPerformance,data.frame-method}
\alias{roc<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{roc} slot}
\usage{
roc(x, ...)

roc(x, ...) <- value

\S4method{roc}{IBRAPerformance}(x)

\S4method{roc}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information necessary to generate ROC curves
for each method and each stratification level.}
}
\value{
The accessor function returns a data frame giving information
  necessary to generate ROC curves for each method and each stratification
  level.
}
\description{
Accessor and replacement functions for the \code{roc} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "roc")
head(roc(ibraperf))
}
\author{
Charlotte Soneson
}

