% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{scatter}
\alias{scatter}
\alias{scatter,IBRAPerformance-method}
\alias{scatter,IBRAPlot-method}
\alias{scatter<-}
\alias{scatter<-,IBRAPerformance,data.frame-method}
\alias{scatter<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'scatter' slot of an IBRAPerformance object}
\usage{
scatter(x, ...)

scatter(x, ...) <- value

\S4method{scatter}{IBRAPerformance}(x)

\S4method{scatter}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'scatter' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

