% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/COBRAData.R
\docType{methods}
\name{score}
\alias{score}
\alias{score<-}
\alias{score,COBRAData-method}
\alias{score<-,COBRAData,data.frame-method}
\title{Accessor and replacement functions for \code{score} slot}
\usage{
score(x, ...)

score(x, ...) <- value

\S4method{score}{COBRAData}(x)

\S4method{score}{COBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{A \code{COBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing scores for each feature and each method.}
}
\value{
The accessor function regurns a data frame containing scores for each
  feature and each method.
}
\description{
Accessor and replacement functions for the \code{score} slot in a
\code{COBRAData} object.
}
\examples{
data(cobradata_example)
head(score(cobradata_example))
}
\author{
Charlotte Soneson
}
