% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{score}
\alias{score}
\alias{score,IBRAData-method}
\alias{score<-}
\alias{score<-,IBRAData,data.frame-method}
\title{Accessors for the 'score' slot of an IBRAData object}
\usage{
score(x, ...)

score(x, ...) <- value

\S4method{score}{IBRAData}(x)

\S4method{score}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAData object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'score' slot of an IBRAData object
}
\author{
Charlotte Soneson
}

