% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{splv}
\alias{splv}
\alias{splv,IBRAPerformance-method}
\alias{splv,IBRAPlot-method}
\alias{splv<-}
\alias{splv<-,IBRAPerformance,character-method}
\alias{splv<-,IBRAPlot,character-method}
\title{Accessor and replacement functions for \code{splv} slot}
\usage{
splv(x, ...)

splv(x, ...) <- value

\S4method{splv}{IBRAPerformance}(x)

\S4method{splv}{IBRAPerformance,character}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A character string giving the name of a feature annotation to
use for stratification.}
}
\value{
The accessor function returns a character string giving the name of a
  feature annotation to use for stratification.
}
\description{
Accessor and replacement functions for the \code{splv} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\examples{
data(ibradata_example)
ibraperf <- calculate_performance(ibradata_example, binary_truth = "status",
                                  aspects = "fdrtpr", splv = "expr_cat")
splv(ibraperf)
}
\author{
Charlotte Soneson
}

