% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{stratiflevels}
\alias{stratiflevels}
\alias{stratiflevels,IBRAPerformance-method}
\alias{stratiflevels,IBRAPlot-method}
\title{Accessors for the stratification levels contained in an IBRAPerformance object}
\usage{
stratiflevels(x, ...)

\S4method{stratiflevels}{IBRAPerformance}(x)
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}
}
\description{
Accessors for the stratification levels contained in an IBRAPerformance object
}
\author{
Charlotte Soneson
}

