% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{tpr}
\alias{tpr}
\alias{tpr,IBRAPerformance-method}
\alias{tpr,IBRAPlot-method}
\alias{tpr<-}
\alias{tpr<-,IBRAPerformance,data.frame-method}
\alias{tpr<-,IBRAPlot,data.frame-method}
\title{Accessors for the 'tpr' slot of an IBRAPerformance object}
\usage{
tpr(x, ...)

tpr(x, ...) <- value

\S4method{tpr}{IBRAPerformance}(x)

\S4method{tpr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An IBRAPerformance object}

\item{...}{Additional arguments}

\item{value}{A data frame}
}
\description{
Accessors for the 'tpr' slot of an IBRAPerformance object
}
\author{
Charlotte Soneson
}

