% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAPerformance.R
\docType{methods}
\name{tpr}
\alias{tpr}
\alias{tpr,IBRAPerformance-method}
\alias{tpr,IBRAPlot-method}
\alias{tpr<-}
\alias{tpr<-,IBRAPerformance,data.frame-method}
\alias{tpr<-,IBRAPlot,data.frame-method}
\title{Accessor and replacement functions for \code{tpr} slot}
\usage{
tpr(x, ...)

tpr(x, ...) <- value

\S4method{tpr}{IBRAPerformance}(x)

\S4method{tpr}{IBRAPerformance,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAPerformance} or \code{IBRAPlot} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame giving information about the observed TPR for each
method and each stratification level, at various adjusted p-value
thresholds.}
}
\description{
Accessor and replacement functions for the \code{tpr} slot in an
\code{IBRAPerformance} or \code{IBRAPlot} object.
}
\author{
Charlotte Soneson
}

