% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/IBRAData.R
\docType{methods}
\name{truth}
\alias{truth}
\alias{truth,IBRAData-method}
\alias{truth<-}
\alias{truth<-,IBRAData,data.frame-method}
\title{Accessor and replacement functions for \code{truth} slot}
\usage{
truth(x, ...)

truth(x, ...) <- value

\S4method{truth}{IBRAData}(x)

\S4method{truth}{IBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{IBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing true assignments and/or scores for
features, together with other feature annotations to use for stratification
of performance calculations.}
}
\description{
Accessor and replacement functions for the \code{truth} slot in an
\code{IBRAData} object.
}
\author{
Charlotte Soneson
}

