% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/COBRAData.R
\docType{methods}
\name{truth}
\alias{truth}
\alias{truth,COBRAData-method}
\alias{truth<-}
\alias{truth<-,COBRAData,data.frame-method}
\title{Accessor and replacement functions for \code{truth} slot}
\usage{
truth(x, ...)

truth(x, ...) <- value

\S4method{truth}{COBRAData}(x)

\S4method{truth}{COBRAData,data.frame}(x) <- value
}
\arguments{
\item{x}{An \code{COBRAData} object.}

\item{...}{Additional arguments.}

\item{value}{A data frame containing true assignments and/or scores for
features, together with other feature annotations to use for stratification
of performance calculations.}
}
\value{
The accessor function returns a data frame containing true
  assignments and/or scores for features, together with other feature
  annotations to use for stratification of performance calculations.
}
\description{
Accessor and replacement functions for the \code{truth} slot in an
\code{COBRAData} object.
}
\examples{
data(cobradata_example)
head(truth(cobradata_example))
}
\author{
Charlotte Soneson
}

